Try and fix a problem in tests that occasionally surfaces

+++ MsbwUtilities/test/Implementation/MsBwUtilityTest/Net/VariableDelimiterTcpClientPlaintextTest.cs
@@ -69,7 +69,10 @@ namespace MsBwUtilityTest.Net
             {
                 _client.Close();
             }
-            _process.Kill();
+            if (!_process.HasExited)
+            {
+                _process.Kill();
+            }
             base.TearDown();
         }
